Stabilize software factory harness for parallel tests
