+++ packages/software-factory/src/harness/shared.ts
@@ -1,0 +1,752 @@
+import {
+  spawnSync,
+  type ChildProcess,
+  type StdioOptions,
+} from 'node:child_process';
+import { createHash } from 'node:crypto';
+import { createServer as createNetServer } from 'node:net';
+import { readdirSync, readFileSync, statSync } from 'node:fs';
+import { join, relative, resolve } from 'node:path';
+
+import jwt from 'jsonwebtoken';
+import '../setup-logger';
+import { logger } from '../logger';
+
+export type RealmAction = 'read' | 'write' | 'realm-owner' | 'assume-user';
+
+export type RealmPermissions = Record<string, RealmAction[]>;
+
+export type FactorySupportContext = {
+  matrixURL: string;
+  matrixRegistrationSecret: string;
+};
+
+export type SynapseInstance = {
+  synapseId: string;
+  port: number;
+  registrationSecret: string;
+};
+
+export interface FactoryRealmOptions {
+  realmDir?: string;
+  realmURL?: URL;
+  realmServerURL?: URL;
+  templateRealmServerURL?: URL;
+  permissions?: RealmPermissions;
+  useCache?: boolean;
+  cacheSalt?: string;
+  templateDatabaseName?: string;
+  context?: FactoryTestContext | FactorySupportContext;
+}
+
+export interface FactoryRealmTemplate {
+  cacheKey: string;
+  templateDatabaseName: string;
+  fixtureHash: string;
+  cacheHit: boolean;
+  realmURL: URL;
+  realmServerURL: URL;
+}
+
+export interface FactoryTestContext extends FactorySupportContext {
+  cacheKey: string;
+  fixtureHash: string;
+  realmDir: string;
+  realmURL: string;
+  realmServerURL: string;
+  templateDatabaseName: string;
+}
+
+export interface StartedFactoryRealm {
+  realmDir: string;
+  realmURL: URL;
+  realmServerURL: URL;
+  databaseName: string;
+  childPids: number[];
+  ports: {
+    publicPort: number;
+    realmServerPort: number;
+    workerManagerPort: number;
+  };
+  cardURL(path: string): string;
+  createBearerToken(user?: string, permissions?: RealmAction[]): string;
+  authorizationHeaders(
+    user?: string,
+    permissions?: RealmAction[],
+  ): Record<string, string>;
+  stop(): Promise<void>;
+}
+
+export type FactoryGlobalContextHandle = {
+  context: FactoryTestContext;
+  stop(): Promise<void>;
+};
+
+export type SpawnedProcess = ChildProcess & {
+  send(message: string): boolean;
+};
+
+export type StartedCompatRealmProxy = {
+  listenPort: number;
+  setTargetPort(targetPort: number): void;
+  stop(): Promise<void>;
+};
+
+export type RunningFactoryStack = {
+  prerender: {
+    stop(): Promise<void>;
+  };
+  realmServer: SpawnedProcess;
+  realmServerURL: URL;
+  workerManager: SpawnedProcess;
+  compatProxy?: StartedCompatRealmProxy;
+  ports: {
+    publicPort: number;
+    realmServerPort: number;
+    workerManagerPort: number;
+  };
+  rootDir: string;
+};
+
+export const packageRoot = resolve(process.cwd());
+export const workspaceRoot = resolve(packageRoot, '..', '..');
+export const realmServerDir = resolve(packageRoot, '..', 'realm-server');
+export const hostDir = resolve(packageRoot, '..', 'host');
+export const baseRealmDir = resolve(packageRoot, '..', 'base');
+export const skillsRealmDir = resolve(
+  packageRoot,
+  '..',
+  'skills-realm',
+  'contents',
+);
+export const sourceRealmDir = resolve(
+  packageRoot,
+  process.env.SOFTWARE_FACTORY_SOURCE_REALM_DIR ?? 'realm',
+);
+export const boxelIconsDir = resolve(packageRoot, '..', 'boxel-icons');
+export const prepareTestPgScript = resolve(
+  realmServerDir,
+  'tests',
+  'scripts',
+  'prepare-test-pg.sh',
+);
+
+export const CACHE_VERSION = 8;
+export const DEFAULT_REALM_SERVER_PORT = Number(
+  process.env.SOFTWARE_FACTORY_REALM_PORT ?? 0,
+);
+export const DEFAULT_COMPAT_REALM_SERVER_PORT = Number(
+  process.env.SOFTWARE_FACTORY_COMPAT_REALM_PORT ?? 0,
+);
+export const DEFAULT_WORKER_MANAGER_PORT = Number(
+  process.env.SOFTWARE_FACTORY_WORKER_MANAGER_PORT ?? 0,
+);
+export const CONFIGURED_REALM_URL = process.env.SOFTWARE_FACTORY_REALM_URL
+  ? new URL(process.env.SOFTWARE_FACTORY_REALM_URL)
+  : undefined;
+export const CONFIGURED_REALM_SERVER_URL = process.env
+  .SOFTWARE_FACTORY_REALM_SERVER_URL
+  ? new URL(process.env.SOFTWARE_FACTORY_REALM_SERVER_URL)
+  : undefined;
+export const DEFAULT_REALM_DIR = resolve(
+  packageRoot,
+  process.env.SOFTWARE_FACTORY_REALM_DIR ?? 'test-fixtures/darkfactory-adopter',
+);
+export const DEFAULT_HOST_URL =
+  process.env.HOST_URL ?? 'http://localhost:4200/';
+export const DEFAULT_ICONS_URL =
+  process.env.ICONS_URL ?? 'http://localhost:4206/';
+export const DEFAULT_ICONS_PROBE_URL = new URL(
+  '@cardstack/boxel-icons/v1/icons/code.js',
+  DEFAULT_ICONS_URL,
+).href;
+export const DEFAULT_PG_PORT = process.env.SOFTWARE_FACTORY_PGPORT ?? '55436';
+export const DEFAULT_PG_HOST =
+  process.env.SOFTWARE_FACTORY_PGHOST ?? '127.0.0.1';
+export const DEFAULT_PG_USER =
+  process.env.SOFTWARE_FACTORY_PGUSER ?? 'postgres';
+export const DEFAULT_PRERENDER_PORT = Number(
+  process.env.SOFTWARE_FACTORY_PRERENDER_PORT ?? 0,
+);
+// The seeded test Postgres used by the harness runs with max_connections=20, so
+// isolated workers need a smaller per-process pool cap to keep workers=2 stable.
+export const DEFAULT_PG_POOL_MAX = Number(
+  process.env.SOFTWARE_FACTORY_PG_POOL_MAX ?? 2,
+);
+export const DEFAULT_MIGRATED_TEMPLATE_DB =
+  process.env.SOFTWARE_FACTORY_MIGRATED_TEMPLATE_DB ??
+  'boxel_migrated_template';
+export const DEFAULT_REALM_LOG_LEVELS =
+  process.env.SOFTWARE_FACTORY_REALM_LOG_LEVELS ??
+  '*=info,realm:requests=warn,realm-index-updater=debug,index-runner=debug,index-perf=debug,index-writer=debug,worker=debug,worker-manager=debug,realm=debug,perf=debug';
+export const DEFAULT_REALM_OWNER = '@software-factory-owner:localhost';
+export const REALM_SECRET_SEED = "shhh! it's a secret";
+export const REALM_SERVER_SECRET_SEED = "mum's the word";
+export const GRAFANA_SECRET = "shhh! it's a secret";
+export const FIXTURE_SOURCE_REALM_URL_PLACEHOLDER = 'https://sf.boxel.test/';
+export const DEFAULT_MATRIX_SERVER_USERNAME =
+  process.env.SOFTWARE_FACTORY_MATRIX_SERVER_USERNAME ?? 'realm_server';
+export const DEFAULT_MATRIX_BROWSER_USERNAME =
+  process.env.SOFTWARE_FACTORY_BROWSER_MATRIX_USERNAME ??
+  'software-factory-browser';
+export const INCLUDE_SKILLS =
+  process.env.SOFTWARE_FACTORY_INCLUDE_SKILLS === '1';
+export const DEFAULT_PERMISSIONS: RealmPermissions = {
+  '*': ['read'],
+  [DEFAULT_REALM_OWNER]: ['read', 'write', 'realm-owner'],
+};
+export const DEFAULT_SOURCE_REALM_PERMISSIONS: RealmPermissions = {
+  '*': ['read'],
+  [DEFAULT_REALM_OWNER]: ['read', 'write', 'realm-owner'],
+};
+export const DEFAULT_BASE_REALM_PERMISSIONS = DEFAULT_SOURCE_REALM_PERMISSIONS;
+export const managedProcessStdio: StdioOptions =
+  process.env.SOFTWARE_FACTORY_DEBUG_SERVER === '1'
+    ? (['ignore', 'inherit', 'inherit', 'ipc'] as const)
+    : (['ignore', 'pipe', 'pipe', 'ipc'] as const);
+export const DEFAULT_REALM_STARTUP_TIMEOUT_MS = Number(
+  process.env.SOFTWARE_FACTORY_REALM_STARTUP_TIMEOUT_MS ?? 120_000,
+);
+export const FULL_INDEX_REALM_STARTUP_TIMEOUT_MS = Number(
+  process.env.SOFTWARE_FACTORY_FULL_INDEX_REALM_STARTUP_TIMEOUT_MS ?? 600_000,
+);
+
+export const harnessLog = logger('software-factory:harness');
+export const supportLog = logger('software-factory:harness:support');
+export const templateLog = logger('software-factory:harness:template');
+export const realmLog = logger('software-factory:harness:realm');
+
+export function formatElapsedMs(elapsedMs: number): string {
+  return `${(elapsedMs / 1000).toFixed(1)}s`;
+}
+
+export async function logTimed<T>(
+  log: ReturnType<typeof logger>,
+  label: string,
+  callback: () => Promise<T>,
+): Promise<T> {
+  let startedAt = Date.now();
+  log.debug(`${label}: starting`);
+  try {
+    let result = await callback();
+    log.debug(
+      `${label}: finished in ${formatElapsedMs(Date.now() - startedAt)}`,
+    );
+    return result;
+  } catch (error) {
+    log.warn(
+      `${label}: failed after ${formatElapsedMs(Date.now() - startedAt)}: ${
+        error instanceof Error ? error.message : String(error)
+      }`,
+    );
+    throw error;
+  }
+}
+
+export function stableStringify(value: unknown): string {
+  if (value === null || typeof value !== 'object') {
+    return JSON.stringify(value);
+  }
+  if (Array.isArray(value)) {
+    return `[${value.map((item) => stableStringify(item)).join(',')}]`;
+  }
+  let record = value as Record<string, unknown>;
+  let keys = Object.keys(record).sort();
+  return `{${keys
+    .map((key) => `${JSON.stringify(key)}:${stableStringify(record[key])}`)
+    .join(',')}}`;
+}
+
+export function hashString(value: string): string {
+  return createHash('sha256').update(value).digest('hex');
+}
+
+export async function findAvailablePort(): Promise<number> {
+  return await new Promise<number>((resolve, reject) => {
+    let server = createNetServer();
+    server.once('error', reject);
+    server.listen(0, '127.0.0.1', () => {
+      let address = server.address();
+      if (!address || typeof address === 'string') {
+        reject(new Error('Unable to determine allocated port'));
+        return;
+      }
+      server.close((error) => {
+        if (error) {
+          reject(error);
+        } else {
+          resolve(address.port);
+        }
+      });
+    });
+  });
+}
+
+export async function resolveFactoryRealmServerURL(
+  realmServerURL?: URL,
+): Promise<URL> {
+  if (realmServerURL) {
+    return new URL(realmServerURL.href);
+  }
+
+  if (CONFIGURED_REALM_SERVER_URL) {
+    return new URL(CONFIGURED_REALM_SERVER_URL.href);
+  }
+
+  let port =
+    DEFAULT_COMPAT_REALM_SERVER_PORT === 0
+      ? await findAvailablePort()
+      : DEFAULT_COMPAT_REALM_SERVER_PORT;
+  return new URL(`http://localhost:${port}/`);
+}
+
+export async function resolveFactoryRealmLocation(options: {
+  realmURL?: URL;
+  realmServerURL?: URL;
+}): Promise<{
+  realmURL: URL;
+  realmServerURL: URL;
+}> {
+  let realmURL = options.realmURL
+    ? new URL(options.realmURL.href)
+    : CONFIGURED_REALM_URL
+      ? new URL(CONFIGURED_REALM_URL.href)
+      : undefined;
+  let realmServerURL = options.realmServerURL
+    ? new URL(options.realmServerURL.href)
+    : CONFIGURED_REALM_SERVER_URL
+      ? new URL(CONFIGURED_REALM_SERVER_URL.href)
+      : undefined;
+
+  if (!realmURL && !realmServerURL) {
+    realmServerURL = await resolveFactoryRealmServerURL();
+    realmURL = new URL('test/', realmServerURL);
+  } else if (!realmServerURL) {
+    throw new Error(
+      'An explicit realm server URL is required when a realm URL is provided. Set options.realmServerURL or SOFTWARE_FACTORY_REALM_SERVER_URL.',
+    );
+  } else if (!realmURL) {
+    realmURL = new URL('test/', realmServerURL);
+  }
+
+  return {
+    realmURL,
+    realmServerURL,
+  };
+}
+
+export function baseRealmURLFor(realmServerURL: URL): URL {
+  return new URL('base/', realmServerURL);
+}
+
+export function skillsRealmURLFor(realmServerURL: URL): URL {
+  return new URL('skills/', realmServerURL);
+}
+
+export function sourceRealmURLFor(realmServerURL: URL): URL {
+  return new URL('software-factory/', realmServerURL);
+}
+
+export function withPort(url: URL, port: number): URL {
+  let next = new URL(url.href);
+  next.port = String(port);
+  return next;
+}
+
+export function realmRelativePath(realmURL: URL, realmServerURL: URL): string {
+  if (realmURL.origin !== realmServerURL.origin) {
+    throw new Error(
+      `Realm URL ${realmURL.href} does not share an origin with realm server URL ${realmServerURL.href}`,
+    );
+  }
+
+  let serverPath = realmServerURL.pathname.endsWith('/')
+    ? realmServerURL.pathname
+    : `${realmServerURL.pathname}/`;
+  if (!realmURL.pathname.startsWith(serverPath)) {
+    throw new Error(
+      `Realm URL ${realmURL.href} is not mounted under realm server URL ${realmServerURL.href}`,
+    );
+  }
+
+  return realmURL.pathname.slice(serverPath.length);
+}
+
+export function realmURLWithinServer(
+  realmServerURL: URL,
+  realmPath: string,
+): URL {
+  return new URL(realmPath || '.', realmServerURL);
+}
+
+export function shouldIgnoreFixturePath(relativePath: string): boolean {
+  if (relativePath === '.DS_Store') {
+    return true;
+  }
+  return relativePath
+    .split('/')
+    .some((segment) =>
+      [
+        'node_modules',
+        '.git',
+        '.boxel-history',
+        'playwright-report',
+        'test-results',
+      ].includes(segment),
+    );
+}
+
+export function hashRealmFixture(realmDir: string): string {
+  let entries: string[] = [];
+
+  function visit(currentDir: string) {
+    for (let entry of readdirSync(currentDir, { withFileTypes: true })) {
+      let absolutePath = join(currentDir, entry.name);
+      let relativePath = relative(realmDir, absolutePath).replace(/\\/g, '/');
+      if (shouldIgnoreFixturePath(relativePath)) {
+        continue;
+      }
+      if (entry.isDirectory()) {
+        visit(absolutePath);
+        continue;
+      }
+      if (!entry.isFile()) {
+        continue;
+      }
+      let stats = statSync(absolutePath);
+      let contentsHash = createHash('sha256')
+        .update(readFileSync(absolutePath))
+        .digest('hex');
+      entries.push(`${relativePath}:${stats.size}:${contentsHash}`);
+    }
+  }
+
+  visit(realmDir);
+  entries.sort();
+  return hashString(entries.join('|'));
+}
+
+export function templateDatabaseNameForCacheKey(cacheKey: string): string {
+  return `sf_tpl_${cacheKey.slice(0, 24)}`;
+}
+
+export function builderDatabaseNameForCacheKey(cacheKey: string): string {
+  return `sf_bld_${cacheKey.slice(0, 16)}`;
+}
+
+export function runtimeDatabaseName(): string {
+  return `sf_run_${Date.now().toString(36)}_${Math.random()
+    .toString(36)
+    .slice(2, 8)}`;
+}
+
+export function pgAdminConnectionConfig(database = 'postgres') {
+  return {
+    host: DEFAULT_PG_HOST,
+    port: Number(DEFAULT_PG_PORT),
+    user: DEFAULT_PG_USER,
+    password: process.env.PGPASSWORD || undefined,
+    database,
+  };
+}
+
+export function quotePgIdentifier(identifier: string): string {
+  if (!/^[a-zA-Z0-9_]+$/.test(identifier)) {
+    throw new Error(`unsafe postgres identifier: ${identifier}`);
+  }
+  return `"${identifier}"`;
+}
+
+export async function waitUntil<T>(
+  condition: () => Promise<T>,
+  options: {
+    timeout?: number;
+    interval?: number;
+    timeoutMessage?: string;
+  } = {},
+): Promise<T> {
+  let timeout = options.timeout ?? 30_000;
+  let interval = options.interval ?? 250;
+  let start = Date.now();
+  while (Date.now() - start < timeout) {
+    let result = await condition();
+    if (result) {
+      return result;
+    }
+    await new Promise((resolve) => setTimeout(resolve, interval));
+  }
+  throw new Error(options.timeoutMessage ?? 'Timed out waiting for condition');
+}
+
+export async function waitForJsonFile<T>(
+  file: string,
+  getLogs: () => string,
+  options: {
+    timeout?: number;
+    label: string;
+    process?: SpawnedProcess;
+  },
+): Promise<T> {
+  let timeout = options.timeout ?? DEFAULT_REALM_STARTUP_TIMEOUT_MS;
+  let startedAt = Date.now();
+
+  while (Date.now() - startedAt < timeout) {
+    try {
+      return JSON.parse(readFileSync(file, 'utf8')) as T;
+    } catch (error) {
+      let nodeError = error as NodeJS.ErrnoException;
+      if (nodeError.code !== 'ENOENT' && !(error instanceof SyntaxError)) {
+        throw error;
+      }
+    }
+
+    if (options.process && options.process.exitCode !== null) {
+      throw new Error(
+        `${options.label} exited early with code ${options.process.exitCode}\n${getLogs()}`,
+      );
+    }
+
+    await new Promise((resolve) => setTimeout(resolve, 100));
+  }
+
+  throw new Error(
+    `Timed out waiting for ${options.label} metadata in ${file}\n${getLogs()}`,
+  );
+}
+
+export function runCommand(command: string, args: string[], cwd: string) {
+  let result = spawnSync(command, args, {
+    cwd,
+    stdio: 'inherit',
+    env: {
+      ...process.env,
+      PGHOST: DEFAULT_PG_HOST,
+      PGPORT: DEFAULT_PG_PORT,
+      PGUSER: DEFAULT_PG_USER,
+    },
+  });
+  if (result.status !== 0) {
+    throw new Error(`command failed: ${command} ${args.join(' ')}`);
+  }
+}
+
+export function cleanupStaleSynapseContainers() {
+  let result = spawnSync(
+    'docker',
+    [
+      'ps',
+      '-aq',
+      '--filter',
+      'name=synapsedocker-',
+      '--filter',
+      'name=boxel-synapse',
+    ],
+    {
+      cwd: workspaceRoot,
+      encoding: 'utf8',
+    },
+  );
+
+  if (result.status !== 0) {
+    return;
+  }
+
+  let containerIds = result.stdout
+    .split(/\s+/)
+    .map((id) => id.trim())
+    .filter(Boolean);
+
+  if (containerIds.length === 0) {
+    return;
+  }
+
+  spawnSync('docker', ['rm', '-f', ...containerIds], {
+    cwd: workspaceRoot,
+    stdio: 'ignore',
+  });
+}
+
+export function maybeRequire(specifier: string) {
+  if (typeof require === 'function') {
+    return require(specifier);
+  }
+  return undefined;
+}
+
+export function fileExists(path: string): boolean {
+  try {
+    return statSync(path).isFile();
+  } catch {
+    return false;
+  }
+}
+
+export function findHostDistPackageDir(): string | undefined {
+  let siblingRoot = resolve(workspaceRoot, '..');
+  let candidates = [
+    process.env.SOFTWARE_FACTORY_HOST_DIST_PACKAGE_DIR,
+    resolve(siblingRoot, 'boxel', 'packages', 'host'),
+    ...readdirSync(siblingRoot, { withFileTypes: true })
+      .filter((entry) => entry.isDirectory())
+      .map((entry) => resolve(siblingRoot, entry.name, 'packages', 'host')),
+    hostDir,
+  ]
+    .filter((value): value is string => Boolean(value))
+    .map((value) => resolve(value));
+
+  let seen = new Set<string>();
+  for (let candidate of candidates) {
+    if (seen.has(candidate)) {
+      continue;
+    }
+    seen.add(candidate);
+
+    if (fileExists(join(candidate, 'dist', 'index.html'))) {
+      return candidate;
+    }
+  }
+
+  return undefined;
+}
+
+export function browserPassword(username: string): string {
+  let cleanUsername = username.replace(/^@/, '').replace(/:.*$/, '');
+  return createHash('sha256')
+    .update(cleanUsername)
+    .update(REALM_SECRET_SEED)
+    .digest('hex');
+}
+
+export function parseFactoryContext(): FactoryTestContext | undefined {
+  let raw = process.env.SOFTWARE_FACTORY_CONTEXT;
+  if (!raw) {
+    return undefined;
+  }
+  return JSON.parse(raw) as FactoryTestContext;
+}
+
+export function hasTemplateDatabaseName(
+  context: FactorySupportContext | FactoryTestContext,
+): context is FactoryTestContext {
+  return 'templateDatabaseName' in context;
+}
+
+export function buildRealmToken(
+  realmURL: URL,
+  realmServerURL: URL,
+  user = DEFAULT_REALM_OWNER,
+  permissions = DEFAULT_PERMISSIONS[DEFAULT_REALM_OWNER] ?? [
+    'read',
+    'write',
+    'realm-owner',
+  ],
+): string {
+  return jwt.sign(
+    {
+      user,
+      realm: realmURL.href,
+      permissions,
+      sessionRoom: `software-factory-session-room-for-${user}`,
+      realmServerURL: realmServerURL.href,
+    },
+    REALM_SECRET_SEED,
+    { expiresIn: '7d' },
+  );
+}
+
+export function createProcessExitPromise(
+  proc: SpawnedProcess,
+  label: string,
+): Promise<never> {
+  return new Promise((_, reject) => {
+    proc.once('exit', (code, signal) => {
+      reject(
+        new Error(
+          `${label} exited before it became ready (code: ${code}, signal: ${signal})`,
+        ),
+      );
+    });
+    proc.once('error', reject);
+  });
+}
+
+export function appendProcessLogs(
+  buffer: string,
+  chunk: Buffer | string,
+): string {
+  return `${buffer}${String(chunk)}`.slice(-100_000);
+}
+
+export function captureProcessLogs(proc: SpawnedProcess) {
+  let stdout = '';
+  let stderr = '';
+
+  proc.stdout?.on('data', (chunk) => {
+    stdout = appendProcessLogs(stdout, chunk);
+  });
+  proc.stderr?.on('data', (chunk) => {
+    stderr = appendProcessLogs(stderr, chunk);
+  });
+
+  return () => {
+    let logs = [];
+    if (stdout) {
+      logs.push(`stdout:\n${stdout}`);
+    }
+    if (stderr) {
+      logs.push(`stderr:\n${stderr}`);
+    }
+    return logs.join('\n\n');
+  };
+}
+
+export async function waitForReady(
+  proc: SpawnedProcess,
+  label: string,
+  timeoutMs = 120_000,
+  getLogs?: () => string,
+): Promise<void> {
+  let timedOut = await Promise.race([
+    new Promise<void>((resolve) => {
+      let onMessage = (message: unknown) => {
+        if (message === 'ready') {
+          proc.off('message', onMessage);
+          resolve();
+        }
+      };
+      proc.on('message', onMessage);
+    }),
+    createProcessExitPromise(proc, label),
+    new Promise<true>((resolve) => setTimeout(() => resolve(true), timeoutMs)),
+  ]);
+
+  if (timedOut) {
+    let logOutput = getLogs?.();
+    throw new Error(
+      `Timed out waiting for ${label} to start${
+        logOutput ? `\n\n${logOutput}` : ''
+      }`,
+    );
+  }
+}
+
+export async function stopManagedProcess(proc: SpawnedProcess): Promise<void> {
+  if (proc.exitCode !== null) {
+    return;
+  }
+  let stopped = new Promise<void>((resolve) => {
+    let onMessage = (message: unknown) => {
+      if (message === 'stopped') {
+        proc.off('message', onMessage);
+        resolve();
+      }
+    };
+    proc.on('message', onMessage);
+  });
+  proc.send('stop');
+  await Promise.race([
+    stopped,
+    new Promise<void>((resolve) => setTimeout(resolve, 15_000)),
+  ]);
+  proc.send('kill');
+}
